reflect(constants): split foreign type impls to their own module
